Refactor: rename pkce package to oidc
